==> src_water_prod/00_preparing_files_for_preprocessing_water_prod.Rmd <==
---
title: "setting up for preprocessing"
author: "Melissa Carew"
date: "2024-02-28"
output: html_document
---
# Preparation for running bioinformatic pipeline
1.  Clone a copy of the workflow from the mecarew github repository https://github.com/mecarew/metabarcoding_workflow into your server directory. You'll need a github account to do this (see - https://docs.github.com/en/get-started/onboarding/getting-started-with-your-github-account).

2.  Make a folder with your miseq run folder for your data in git folder on the uomshare, then transfer data files from the AGRF server onto into your git folder on the uomshare in a folder called 'raw' (this is the folder name which these raw data files are accessed in the workflow)
 
3.    Format a sample.tsv file using the script below. This will be required to run the 01_preprocessing.

```{r}
library(dplyr)

# set up metabarcoding workflow folder inside MiSeq run file for input and output data files (NOTE each new project will require the directories to be setup for your dataset, this code has input data in the folder 'test')

#miseq_folder <- "miseq21"

# Set the directory path for files to be stored on uomshare (NOTE: 'test' will need to be change to your miseq folder name

uomshare <- paste0("~/uomShare/wergStaff/MelCarew/git-data/metabarcoding_workflow/",miseq_folder,"/metabarcoding_workflow/")

if (dir.exists(uomshare)) {
  warning("Directory already exists. Please change the name to the name of your Miseq folder. e.g. ~/uomShare/wergStaff/MelCarew/git-data/miseqXX/metabarcoding_workflow/ This will need to be done for each new miseq dataset at the begining of each notebook")
}

dir.create(uomshare)
```

```{r}
# create data directory for data analysis
data <- paste0(uomshare,"data")
dir.create(data)

#create data folder for sample information need for workflow
metadata <- paste0(uomshare,"data/metadata")
dir.create(metadata)

# move raw data into the 'metabarcoding workflow' folder for analysis
file.rename(
  from = paste0("~/uomShare/wergStaff/MelCarew/git-data/metabarcoding_workflow/", miseq_folder, "/raw"), 
  to = paste0(uomshare, "data/raw")
)
# make the mteadata.tsv file for analysis
# list files in the folder containing raw data files
file_names <- list.files(paste0(uomshare, "data/raw"))

# create a dataframe with the file names
sample_list <- data.frame(all_files= file_names)

# split columns for R1 and R2 files
samples <- sample_list %>%
  dplyr::mutate(
    filename_r1  = ifelse(grepl("_R2", all_files), all_files, NA),
    filename_r2 = ifelse(grepl("_R1", all_files), all_files, NA)
  )

samples$filename_r1 <- ifelse(is.na(samples$filename_r1), sample_list$all_files, NA)
samples$filename_r2 <- ifelse(is.na(samples$filename_r2), sample_list$all_files, NA)

# remove 'all_files' column
samples <- samples[, -which(names(samples) == "all_files")]

# remove gaps
samples$filename_r1 <- na.omit(samples$filename_r1)
samples$filename_r2 <- na.omit(samples$filename_r2)

# remove duplicates
samples <- unique(samples)

# add 'sample names' column
samples$sample_name <- substr(samples$filename_r1, 1, regexpr("_", samples$filename_r1) - 1)

# add 'factor' column
samples$factor <- substr(samples$sample_name, 1, nchar(samples$sample_name))
# remove replicate data
samples <- samples %>%
   dplyr::mutate(factor = sub("(rep1|rep2|rep3|rep4|rep5|rep6)$", "", factor))

# Add directory "/data/raw/" to the beginning of the "File_Name" column
samples$filename_r1 <- paste0(uomshare, "data/raw/",samples$filename_r1)
samples$filename_r2 <- paste0(uomshare, "data/raw/",samples$filename_r2)

# Write table '.tsv' to analysis directory
write.table(samples, file = paste0(uomshare, "data/metadata/samples.tsv"), sep = "\t", row.names = FALSE, quote = FALSE)

# Check that your a file called 'samples.tsv' is in your 'data' folder.
```

# Some more on using github (if you haven't before)
Git can be added to the top right pane of your rstudio window using the instructions below:

Configure Git in RStudio:
Open RStudio and go to the "Tools" menu. Choose "Global Options" and then click on the "Git/SVN" tab. Here, you need to specify the path to the Git executable. If Git is installed in the default location, RStudio should automatically detect it. Otherwise, you may need to browse and select the Git executable manually.
