==> src_water_prod/05_final_summary_water_prod.Rmd <==
---
title: "final_data_summary"
author: "Jessica Chung /Melissa Carew"
date: "01/03/2024"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This final section enables data to be arranged into a final table for fileing and filtering

# producing final data summary

Starting from here will need the threshold list and the sample_df from Cache03
```{r}
#reset directory 
#***NOTE: YOU WILL NEED TO CHANGE THE MISEQ FOLDER NAME to the directory you are working in***
miseq_folder <- "miseq20"

# Set the directory path for files to be stored on uomshare (NOTE: 'test' will need to be change to your miseq folder name)
uomshare <- paste0("~/uomShare/wergStaff/MelCarew/git-data/metabarcoding_workflow/",miseq_folder,"/metabarcoding_workflow")

#uomshare <- paste0("~/uomShare/wergStaff/MelCarew/git-data/",miseq_folder,"/metabarcoding_workflow")

# load thresholds_list form cahche03 
thresholds_list <- readRDS(paste0(uomshare,"/cache/thresholds_list.rds"))
samples_df <- readRDS(paste0(uomshare,"/cache/samples_df.rds"))

# load packages
library(dplyr)
library(tidyverse)
```

# load results files from previuos steps
```{r}
short_lulu <- readRDS(paste0(uomshare,"/lulu/short_lulu.rds"))
right_lulu <- readRDS(paste0(uomshare,"/lulu/right_lulu.rds"))
short_classifications <- qiime2R::read_qza(paste0(uomshare,"/classification/short_classification.qza"))
right_classifications <- qiime2R::read_qza(paste0(uomshare,"/classification/right_classification.qza"))
short_seq <- qiime2R::read_qza(paste0(uomshare,"/dada2/short_representative_sequences.qza"))
right_seq <- qiime2R::read_qza(paste0(uomshare,"/dada2/right_representative_sequences.qza"))
```

# check classification data for short amplicon
```{r}
short_classifications$data %>% head(100)
```

# check classification data for right amplicon
```{r}
right_classifications$data %>% head(100)
```


```{r}
# Pull out table of read counts with LULU filter applied
short_lulu$curated_table
```


```{r}
# Pull out table of read counts with LULU filter applied
right_lulu$curated_table
```

```{r}
# merge short_classifications with tHersholds
short_df <- merge(short_classifications$data, thresholds_list[["short"]], by.x="Feature.ID", by.y="feature-id")

right_df <- merge(right_classifications$data, thresholds_list[["right"]], by.x="Feature.ID", by.y="feature-id")
```

```{r}
short_seq$data %>% head
```

```{r}
right_seq$data %>% head
```

```{r}
short_df %>% arrange(desc(max_p_identity)) %>% head(100)
```

```{r}
right_df %>% arrange(desc(max_p_identity)) %>% head(100)
```

```{r}
table(short_df$Taxon == "Unassigned")
```

```{r}
table(right_df$Taxon == "Unassigned")
```

```{r}
short_df %>% filter(Taxon == "Unassigned") %>% pull(max_p_identity) %>% table
```

```{r}
right_df %>% filter(Taxon == "Unassigned") %>% pull(max_p_identity) %>% table
```


```{r}
# for big dataset
short_search_results <- readr::read_tsv(paste0(uomshare,"/classification/preliminary/short_vsearch.tsv"))
```

```{r}
short_search_results$data %>% dim
```

```{r}
# for big dataset
right_search_results <- readr::read_tsv(paste0(uomshare,"/classification/preliminary/right_vsearch.tsv"))
```

## Combine workflow outputs

# fix sample names (if they have 'X' at the start)
This is the case for sampled from the 10 sites study as the names start with numbers

```{r}
all_tables <- list(
  short=short_lulu$curated_table,
  right=right_lulu$curated_table
)
```


```{r}
# Get sample names and groups
#sample_groups <- samples_df$factor
#names(sample_groups) <- samples_df$sample_name
#sample_groups
```

# combining read tables for the short and right amplicons
```{r}
#make them dataframes
short_read_tab <- all_tables$short
right_read_tab <- all_tables$right
```

```{r}
# Do the amplicon datasets contain the same number of samples?
short_read_tab %>% dim
right_read_tab %>% dim

# are these samples the same for both amplicons?
if(identical(names(short_read_tab), names(right_read_tab))) {
  print("Column names match.")
} else {
  print("Column names do not match.")
}
```

# if column names match skip this step.
```{r}
# if they do not match which columns are different?
names_only_in_short_read_tab <- setdiff(names(short_read_tab), names(right_read_tab))
print("Column names present only in short_read_tab:")
print(names_only_in_short_read_tab)

names_only_in_right_read_tab <- setdiff(names(right_read_tab), names(short_read_tab))
print("Column names present only in right_read_tab:")
print(names_only_in_right_read_tab)
```

# write scripts to add a blank column to align data for all amplicons
```{r}
# align data set by adding blank column for columns not present in one amplicon but not the other (Cont3rep1, Cont1rep2). 
# short_read_tab <- mutate(short_read_tab, Cont3rep1 = c(0))
# right_read_tab <- mutate(right_read_tab, Cont1rep2 = c(0))
# right_read_tab <- dplyr::mutate(right_read_tab, CO1Cont2 = c(0))
# right_read_tab <- dplyr::mutate(right_read_tab, CO1Cont3 = c(0))
# right_read_tab <- dplyr::mutate(right_read_tab, COICont1 = c(0))
```

```{r}
#merge short and right reads tables into one dataframe
all_read_df <-  rbind(short_read_tab, right_read_tab)

# covert row headers to column and then name column 1
all_read_df <- tibble::rownames_to_column(all_read_df, var = "asv_code")
```

# combining read classification tables for the short and right amplicons
```{r}
# bind short and right classification data
all_classif_df <-  rbind(short_df, right_df)

# rename column 1
colnames(all_classif_df)[1] <- "asv_code"

# join into a singe dataframe
all_readclass_df <- left_join(all_classif_df, all_read_df, by = "asv_code")
```

# add in ASV sequence data
```{r}
# put sequences in a dataframe
short_asvseq <- data.frame(short_seq$data)
right_asvseq <- data.frame(right_seq$data)

# covert row headers to column and then name column 1
short_asvseq <- tibble::rownames_to_column(short_asvseq, var = "asv_code")
right_asvseq <- tibble::rownames_to_column(right_asvseq, var = "asv_code")

# add amplicon information
short_asvseq  <- mutate(short_asvseq, amplicon = "short")
right_asvseq <- mutate(right_asvseq, amplicon = "right")

# align column names for merging
short_asvseq <- short_asvseq %>% dplyr::rename(asv_seq = short_seq.data)
right_asvseq <- right_asvseq %>% dplyr::rename(asv_seq = right_seq.data)


# merge data into a single dataframe
all_asvseq <-  rbind(short_asvseq, right_asvseq)
```

```{r}
# merge ALL data into a single dataframe
final_df <- left_join(all_readclass_df, all_asvseq, by = "asv_code")

# remove columns without data
final_df <- na.omit(final_df)

final_df$Taxon <- gsub("k__", "", final_df$Taxon)
final_df$Taxon <- gsub("p__", "", final_df$Taxon)
final_df$Taxon <- gsub("c__", "", final_df$Taxon)
final_df$Taxon <- gsub("o__", "", final_df$Taxon)
final_df$Taxon <- gsub("f__", "", final_df$Taxon)
final_df$Taxon <- gsub("g__", "", final_df$Taxon)
final_df$Taxon <- gsub("s__", "", final_df$Taxon)

# place taxonomic data in separate columns
final_df <- tidyr::separate(final_df, Taxon, into = c("kingdom", "phylum", "class", "order", "family", "genus","species"), sep = ";",  fill = "right")

# final tidy up (remove 'NA' and X from start of sample names)
final_df[is.na(final_df)] <- ""
head(final_df)
```
```{r}
# fix taxonomic discrepancies between BOLD and GenBank
final_df <- final_df %>%
  mutate(
    class = gsub("Copeopda", "Hexanauplia", class),    # Replace 'Copeopda' with 'Hexanauplia'
    order = gsub("Tubificida", "Haplotaxida", order),   # Replace 'Tubificida' with 'Haplotaxida'
     order = gsub("Oomycota", "", order)
   )
# Move 'Oomycota' from 'phylum' to 'class' and replace 'Oomycota' with 'Heterokontophyta' in 'phylum'
final_df<- final_df %>%
  mutate(
    class = ifelse(phylum == "Oomycota", "Oomycota", class),              # Move 'Oomycota' to 'class'
    phylum = ifelse(phylum == "Oomycota", "Heterokontophyta", phylum)     # Replace 'Oomycota' with 'Heterokontophyta'
  )
# Move 'Oomycota' from 'phylum' to 'class' and replace 'Oomycota' with 'Heterokontophyta' in 'phylum'
final_df <- final_df %>%
  mutate(
    class = ifelse(phylum == "Discosea", "Discosea", class),              # Move 'Oomycota' to 'class'
    phylum = ifelse(phylum == "Discosea", "Amoebozoa", phylum)     # Replace 'Oomycota' with 'Heterokontophyta'
  )

final_df <- final_df %>%
  mutate(
    family = gsub("Oceaniidae", "Cordylophoridae", family),   # Replace 'Oceaniidae' with 'Cordylophoridae'
    family = gsub("Telephlebiidae", "Aeshnidae", family))  

```


# save raw data from 
```{r}
# save dataframe create a directory for results
dir.create(paste0(uomshare, "/final_data_summaries"))

timestamp <- format(Sys.time(), "%d_%m_%Y")

# Create the file name with the timestamp 
file_name <- paste0(uomshare, "/final_data_summaries/", miseq_folder, "_vsearch_data_summary_", timestamp, ".csv")
# Save the dataframe with the new file name
write.csv(final_df, file = file_name, row.names = FALSE)
```
